requirements.txt
django
pillow
mysqlclient
django-filter
python-dotenv
django-spurl
djangorestframework